ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.3
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0
datasets==2.14.7
accelerate==0.24.1
sentencepiece==0.1.99
protobuf==4.25.1